README.md
This program displays a small GUI to load and select .CSV files generated by JASCO spectrometers. 
Up to six spectra can be overlayed and individual line colors can be changed in a color picker. 
The displayed range can be adjusted both in x and in y. Additionally, a button sets the minimum of the curve = 0.
"Create PDF output" generates a pdf with all the loaded curves and names assigned, plotted in matplotlib.